refactor: reorganize benchmark parameters add each parameter.py in reads , writes folder
